change: move AP Calculator into own page
