--- src/components/APCalcCard/index.js
@@ -1,2 +1,0 @@
-import APCalcCard from "./APCalcCard";
-export default APCalcCard;
