+++ src/components/ApCalc/Result.jsx
