+++ src/utils/navRoutes.js
@@ -1,4 +1,5 @@
 export const navRoutes = [
   { link: "/", text: "Timers", test: /^\/$/ },
-  { link: "/upgrades", text: "Upgrades" }
+  { link: "/upgrades", text: "Upgrades" },
+  { link: "/ap-calc", text: "AP Calculator" }
 ];
